Provide mainActivityAlias
