+++ common/src/main/kotlin/com/github/kr328/clash/common/di/AppInfoProvider.kt
@@ -1,5 +1,6 @@
 package com.github.kr328.clash.common.di
 
+import android.content.ComponentName
 import kotlin.reflect.KClass
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.get
@@ -10,6 +11,7 @@ interface AppInfoProvider {
   val receiveBroadcastsPermission: String
   val mainActivityClass: KClass<*>
   val restartReceiverClass: KClass<*>
+  val mainActivityAlias: ComponentName
 
   companion object : KoinComponent {
     val appInfoProvider: AppInfoProvider = get()
