+++ common/src/main/kotlin/com/github/kr328/clash/common/util/Intent.kt
@@ -1,12 +1,8 @@
 package com.github.kr328.clash.common.util
 
-import android.content.ComponentName
-import android.content.Context
 import android.content.Intent
-import android.content.pm.PackageManager
 import android.net.Uri
 import android.os.Build
-import com.github.kr328.clash.common.compat.queryIntentActivitiesCompat
 import com.github.kr328.clash.common.di.AppInfoProvider.Companion.appInfoProvider
 import java.io.Serializable
 import kotlin.uuid.Uuid
@@ -15,25 +11,6 @@ fun mainIntent(block: Intent.() -> Unit = {}): Intent {
   val mainActivityClass = appInfoProvider.mainActivityClass
   return mainActivityClass.intent.apply(block = block)
 }
-
-val Context.mainActivityAlias: ComponentName
-  get() {
-    val mainActivityName = appInfoProvider.mainActivityClass.java.name
-    val launcherIntent = Intent(Intent.ACTION_MAIN).addCategory(Intent.CATEGORY_LAUNCHER)
-    val resolveFlags = PackageManager.MATCH_DISABLED_COMPONENTS
-
-    return packageManager
-      .queryIntentActivitiesCompat(launcherIntent, resolveFlags)
-      .firstNotNullOfOrNull { resolveInfo ->
-        val activityInfo = resolveInfo.activityInfo
-        if (activityInfo.targetActivity == mainActivityName) {
-          ComponentName(activityInfo.packageName, activityInfo.name)
-        } else {
-          null
-        }
-      }
-      ?: error("Launcher alias targeting $mainActivityName is not declared in AndroidManifest.xml")
-  }
 
 fun Intent.grantPermissions(read: Boolean = true, write: Boolean = true): Intent = apply {
   var flags = 0
